+++ src/lib/utils/array.ts
@@ -3,12 +3,29 @@
  * the item will be inserted later will be placed earlier in the array.
  * @param arr modified by inserting item.
  * @param item
+ * @deprecated this is confusing, it sorts with lower priority being placed earlier. Prefer insertOrderSorted, which is nearly the same.
  */
 export function insertPrioritySorted<T extends { priority: number }>(
     arr: T[],
     item: T
 ): T[] {
     const index = binaryFindPartition(arr, (v) => v.priority >= item.priority);
+    arr.splice(index === -1 ? arr.length : index, 0, item);
+    return arr;
+}
+
+/**
+ * Inserts an item into an array sorted by order. If two items have the same order,
+ * the item inserted later will be placed later in the array.
+ * The array will be sorted with lower order being placed sooner.
+ * @param arr modified by inserting item.
+ * @param item
+ */
+export function insertOrderSorted<T extends { order: number }>(
+    arr: T[],
+    item: T
+): T[] {
+    const index = binaryFindPartition(arr, (v) => v.order > item.order);
     arr.splice(index === -1 ? arr.length : index, 0, item);
     return arr;
 }