+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Actually fixed `@category` tag incorrectly appearing on function types if used on a type alias, #1745.
+-   Fix error in console when a page contains no documentation items.
 
 ## v0.22.7 (2021-10-25)